auto loading of typedefs and resolvers
